# backend/generated/openapi/src/main/java/com/whipup/generated/api/RecipesApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.19.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.whipup.generated.api;

import com.whipup.generated.model.ErrorResponse;
import com.whipup.generated.model.RecipeDetailResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-09-23T01:37:49.076375542Z[Etc/UTC]", comments = "Generator version: 7.19.0")
@Validated
@Tag(name = "Recipe", description = "the Recipe API")
public interface RecipesApi {

    default RecipesApiDelegate getDelegate() {
        return new RecipesApiDelegate() {};
    }

    String PATH_GET_RECIPE_DETAIL = "/recipes/{recipeId}";
    /**
     * GET /recipes/{recipeId}
     *
     * @param recipeId  (required)
     * @return Recipe detail (status code 200)
     *         or Authentication failed or missing (status code 401)
     *         or Resource not found (status code 404)
     */
    @Operation(
        operationId = "getRecipeDetail",
        tags = { "Recipe" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Recipe detail", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RecipeDetailResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentication failed or missing", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = RecipesApi.PATH_GET_RECIPE_DETAIL,
        produces = { "application/json" }
    )
    default ResponseEntity<RecipeDetailResponse> getRecipeDetail(
        @NotNull @Parameter(name = "recipeId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("recipeId") Long recipeId
    ) {
        return getDelegate().getRecipeDetail(recipeId);
    }

}