# backend/generated/openapi/src/main/java/com/whipup/generated/api/RecommendationsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.19.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.whipup.generated.api;

import com.whipup.generated.model.ErrorResponse;
import org.springframework.lang.Nullable;
import com.whipup.generated.model.RecommendationPage;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-09-23T01:37:49.076375542Z[Etc/UTC]", comments = "Generator version: 7.19.0")
@Validated
@Tag(name = "Recommendation", description = "the Recommendation API")
public interface RecommendationsApi {

    default RecommendationsApiDelegate getDelegate() {
        return new RecommendationsApiDelegate() {};
    }

    String PATH_GET_RECOMMENDATIONS = "/recommendations";
    /**
     * GET /recommendations
     *
     * @param missingCount  (required)
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 30)
     * @return Recommendations (status code 200)
     *         or Invalid request (status code 400)
     *         or Authentication failed or missing (status code 401)
     *         or Request conflicts with current state (status code 409)
     */
    @Operation(
        operationId = "getRecommendations",
        tags = { "Recommendation" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Recommendations", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RecommendationPage.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentication failed or missing", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "Request conflicts with current state", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = RecommendationsApi.PATH_GET_RECOMMENDATIONS,
        produces = { "application/json" }
    )
    default ResponseEntity<RecommendationPage> getRecommendations(
        @NotNull @Parameter(name = "missingCount", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "missingCount", required = true) Integer missingCount,
        @Min(value = 0) @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Min(value = 1) @Max(value = 100) @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "30") Integer size
    ) {
        return getDelegate().getRecommendations(missingCount, page, size);
    }

}